% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{bowen_map_ggplot}
\alias{bowen_map_ggplot}
\title{Bowen Map Plotting (Improved)}
\usage{
bowen_map_ggplot(
  gg_func,
  title = "Default",
  subtitle = "Default",
  caption = "Default",
  pal = "ag_GrnYl"
)
}
\arguments{
\item{gg_func}{function, that returns ggplot components}

\item{title}{Text, title for map}

\item{subtitle}{Text, subtitle for map}

\item{caption}{Text, caption for map}

\item{pal}{Palette, from colorspace package}
}
\value{
ggplot
}
\description{
This function takes aas an input a function that returns ggplot components
(See: https://ggplot2-book.org/programming.html). This allows the
bowen_map_ggplot function to be able to add ggplot components before and
after the input layers. This is useful as this allows additional map
elements, such as road / trails, to be added above the main feature layers.
}
