% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sdm.R
\name{invert}
\alias{invert}
\title{Invert raster, so max values are min and vice versa.}
\usage{
invert(input_rast)
}
\arguments{
\item{input_rast}{input raster}
}
\value{
terra rast that has been inverted
}
\description{
Invert raster, so max values are min and vice versa.
}
